Intial pyinotify test program
The program is called monitor.py. It will monitor the directory monitor_me. Whenever a text file is created (it has to be empty) in the monitored directory it will have two random strings placed in it. The first string will be the user name and the second string will be a password. The strings are delimeted with a / character. For easy testing and a way to keep our monitor_me directory clean I wrote a quick script. I have yet to daemonize this.

SCRIPT USAGE: (1) cd test-dev/python-pyinotify
	      (2) python monitor.py
	      (3) open a new bash window and cd into test-dev/python-pyinotify
	      (4) type ./test.sh
Read the script to see what it does. It is very simple.