Added a script to help create db's
It was a little bit of a hassle setting up a test database for the test.py program.
So I found this nice bash script on the googlenet. It is self explanatory.
Just cd into the python-MySQLdb directory and type ./create_db.sh.
It will start taking user input at that point.

For the test.py program use the script as follow:
your root password
database will be called, TEST (caps matter)
user will be named , test
password will be, test123