Foreground/Daemon run/log modes available
Implemented 2 new functions and changed a third:
	log(tag, message) -- this function is used for logging non-error
	messages.

	logError(tag, message) -- this function is used for logging errors.

	error(tag, e) used to be printError(e). -- this function calls the
	logError function with parameters based on the value's recieve from
	the exception object passed in.
These functions are used in approriate places. I haven't tested this
very robustly yet. But the service still runs and works as before.

This commit closes #5. I had to change the logging system around a bit
to compliment both modes we can run our service in.