The service now creates a pid file
The file is at /var/run/qwifi.pid.lock
Currently if we try to launch the service a second time it will run but
the second instance will sit and do nothing until the first instance is
done or killed. I will create a new issue for this since I don't have
time to finish it right now. This commit closes #8